+++ src/models/signature.rs
@@ -3,7 +3,7 @@ use serde::{Serialize, Deserialize};
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Signature {
-    pub v: u8,
+    pub v: u64,
     pub r: String,
     pub s: String
 }
